requirements.txt
accelerate
aiofiles==23.2.1
aiohttp
aiosignal
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.0
async-timeout
attrs
blinker
boltons
Brotli
certifi==2023.7.22
cffi
charset-normalizer
click
colorama
conda==4.3.16
conda-libmamba-solver
conda-package-handling
conda_package_streaming
contourpy==1.3.0
cryptography
cycler==0.12.1
dataclasses
datasets
diffusers
dill
einops==0.6.1
exceptiongroup==1.2.2
fastapi==0.115.0
ffmpy==0.4.0
filelock
firecrawl==1.2.4
Flask
fonttools==4.54.1
frozenlist
fsspec
gmpy2
gradio==4.44.1
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
idna
importlib-metadata
importlib_resources==6.4.5
invisible-watermark==0.2.0
itsdangerous
Jinja2
joblib
jsonpatch
jsonpointer==2.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
libmambapy
mamba
markdown-it-py==3.0.0
MarkupSafe
matplotlib==3.9.2
mdurl==0.1.2
mpmath
multidict
multiprocess
narwhals==1.9.1
nest-asyncio==1.6.0
networkx
numpy
omegaconf==2.3.0
opencv-python==4.8.0.76
orjson==3.10.7
packaging
pandas
Pillow
pluggy
psutil
pyarrow==12.0.1
pycosat
pycparser
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyOpenSSL
pyparsing==3.1.4
PySocks
python-dateutil
python-dotenv==1.0.1
python-multipart==0.0.12
pytz
PyWavelets==1.4.1
PyYAML
referencing==0.35.1
regex
requests
rich==13.9.2
rpds-py==0.20.0
ruamel.yaml
ruamel.yaml.clib
ruff==0.6.9
sacremoses
safetensors
scipy==1.11.2
semantic-version==2.10.0
shellingham==1.5.4
six
sniffio==1.3.1
starlette==0.38.6
sympy
tokenizers
tomlkit==0.12.0
toolz
torch
torchsde==0.2.5
tqdm
trampoline==0.1.2
transformers
typer==0.12.5
typing_extensions==4.12.2
tzdata
urllib3
uvicorn==0.31.0
websockets==11.0.3
Werkzeug
xxhash
yarl
zipp
zstandard==0.19.0